Requirements
Write a program which reads a numbers text file.  The file contains a series of numbers, each written on a separate line.  The program will read the contents of the file into an array and then display the following information:

the lowest number in the array
the highest number in the array
the total of the numbers in the array
the average of the numbers in the array.